customflags: Move Size from ucs to customflags
